Move function to static method in a modflow mixin
